Replaced the trajectory constructor and validation in the cpp file
